backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
biopython>=1.84
numpy>=1.26.0
